==> 0.packages.r <==
#testtextforcommit
#https://cran.r-project.org/web/packages/stm/vignettes/stmVignette.pdf
library(languageserver)
#library(httpgd)
library(lda)
library(slam)
library(stm)
library(tidyverse)
library(tidytext)
library(quanteda)
library(stm)
library(stopwords)
#library(qdap)
library(dplyr)
#stop_words = stopwords(language = "sv",source = "nltk")

#STM prepare associating text with metadata

library(tm)
#library(txtorg)
library(slam)
library(Matrix)

#Tidy STM and GGplot
library(ggplot2)
library(tidystm)

# 
library("stringr")
library("dplyr")
library("kableExtra")
library("phantomjs")
library("magrittr")
library("magick")
library("glue")

#
library(stm)
library(furrr)

#Topic model k0
library(geometry)
library(rsvd)
library(Rtsne)

#Packages Section 1: Plot Topic Themes
library(ggthemes)
library(scales)
library(tibble)

library(lubridate)

#5 estimate effect

#7 describe
library(foreach)
library(doParallel)

#visualisations

#2.2 processbyyear
#load(file="/home/petter/data/ThesisCode/Topic Model/RData/dataParagraph.Rdata")
library("imputeTS")
library(ggplot2)
library(datetime)
library(dplyr)

library(tidyverse)
library(lubridate)
library(timetk)
library(zoo)
library(xts)

library(plotly)
library(tidyverse)
library(tidyquant)
library(ggplot2)
library(lubridate)

library(hrbrthemes)

#visualisations
library(fuzzyjoin)
library(sqldf)

library(reshape2)
library(mudata)

library(purrr)
library(furrr)

library(gt)
library(gtsummary)

library(ggstream)

library(changepoint)
library(arsenal)
